refactor: consolidate event and timeline models into a single models file
